requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.1
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
pydantic[email]==2.1.1
python-dotenv==1.0.0
pytest==7.4.0
docutils
python-jose[cryptography]
asyncpg>=0.27.0
passlib[bcrypt]
numpy